# exp/orderbook: Represent assets in orderbook graph as int32 instead of strings