--- a/services/horizon/internal/paths/main.go
+++ b/services/horizon/internal/paths/main.go
@@ -20,16 +20,16 @@ type Query struct {
 
 // Path is the result returned by a path finder and is tied to the DestinationAmount used in the input query
 type Path struct {
-	Path              []xdr.Asset
-	Source            xdr.Asset
+	Path              []string
+	Source            string
 	SourceAmount      xdr.Int64
-	Destination       xdr.Asset
+	Destination       string
 	DestinationAmount xdr.Int64
 }
 
 // Finder finds paths.
 type Finder interface {
-	// Return a list of payment paths and the most recent ledger
+	// Find returns a list of payment paths and the most recent ledger
 	// for a Query of a maximum length `maxLength`. The payment paths
 	// are accurate and consistent with the returned ledger sequence number
 	Find(ctx context.Context, q Query, maxLength uint) ([]Path, uint32, error)